Backend/pyproject.toml
[project]
name = "InstaShare-backend"
version = "0.1.0"
description = "InstaShare's backend: Online platform for managing documents and share its with FastAPI."
authors = [
    { name = " InstaShare's team", email = "inoid2007@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase",
    "alembic"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
